Kang-il/servlet_test_server#10 2021 07 07 JSP 탬플릿과 DB 연동 (수정)

+++ src/main/java/com/test/lesson04/Ex02Delete.java
@@ -10,26 +10,27 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import com.test.common.MysqlService;
-
+ 
 @WebServlet("/lesson04/ex02_delete")
 public class Ex02Delete extends HttpServlet{
 	@Override
 	public void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		int id =Integer.parseInt(request.getParameter("id"));
+		Integer id =Integer.parseInt(request.getParameter("id"));
 		
-		
-		MysqlService mysqlService=MysqlService.getInstance();
-		String query="DELETE FROM `new_user` WHERE id="+id;
-		mysqlService.connection();
-		
-		try {
-			mysqlService.update(query);
-		} catch (SQLException e) {e.printStackTrace();
-		} finally {
-			mysqlService.disconnect();
+		if(id!=null) {
+			MysqlService mysqlService=MysqlService.getInstance();
+			String query="DELETE FROM `new_user` WHERE id="+id;
+			mysqlService.connection();
+			
+			try {
+				mysqlService.update(query);
+			} catch (SQLException e) {e.printStackTrace();
+			} finally {
+				mysqlService.disconnect();
+			}
+			
+			
+			response.sendRedirect("/lesson04/ex02/ex02_1.jsp");
 		}
-		
-		
-		response.sendRedirect("/lesson04/ex02/ex02_1.jsp");
 	}
 }
